requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
decorator==5.1.1
Django==3.2.12
django-extensions==3.1.5
djangorestframework==3.13.1
executing==0.8.3
ipython==8.2.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.28
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
wcwidth==0.2.5
